Add 2 challenges, fix 2 challenges, qol updates
This adds two new challenges based exploiting on the global config
pattern and introducing the `HTMLCollection` clobbering technique.

It also fixes a bug in the two `<form>` clobbering challenges which
did not account for solutions using the `id` attribute instead of the
`name` attribute.

As a quality of life (qol) update, I added a home and previous button
to each challenge page for improve navigation.

Lastly, this also fixed a bug in the `encode.js` script where it would
output an invalid script if the key is longer than the value. If this
happens to be the case we should just ignore the rest of the key
(whereas previously it would extend the value to match the key length).